Restructure repository for better splitting
